# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Advanced Talent Agency - Sports Streaming & Betting Exchange
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List games
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: sport
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameListResponse"
    post:
      operationId: createGame
      tags: [games]
      summary: Create game (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update game (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/settle:
    post:
      operationId: settleGame
      tags: [games]
      summary: Settle game result (admin/moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettleGameInput"
      responses:
        "200":
          description: Settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/cancel:
    post:
      operationId: cancelGame
      tags: [games]
      summary: Cancel game and refund all bets (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/bets:
    get:
      operationId: listGameBets
      tags: [games]
      summary: List all bets for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bet"

  # BETS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List my notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /reports/streaming:
    get:
      operationId: getStreamingReport
      tags: [reports]
      summary: Streaming revenue report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
            default: monthly
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamingReport"
  /reports/betting:
    get:
      operationId: getBettingReport
      tags: [reports]
      summary: Betting/brokerage revenue report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
            default: monthly
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BettingReport"
  /reports/wallets:
    get:
      operationId: getWalletReport
      tags: [reports]
      summary: Wallet transaction report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
            default: monthly
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletReport"
components:
  schemas:
...
    Game:
      type: object
      required: [id, sport, playerA, playerB, eventDate, eventTime, status, createdAt]
      properties:
        id:
          type: integer
        sport:
          type: string
          enum: [pool, boxing]
        playerA:
          type: string
        playerB:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed, cancelled]
        result:
          type: ["string", "null"]
          enum: [player_a_wins, player_b_wins, draw, null]
        totalBetPool:
          type: number
          default: 0
        openBetsCount:
          type: integer
          default: 0
        matchedBetsCount:
          type: integer
          default: 0
        createdAt:
          type: string
    GameInput:
      type: object
      required: [sport, playerA, playerB, eventDate, eventTime]
      properties:
        sport:
          type: string
          enum: [pool, boxing]
        playerA:
          type: string
        playerB:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
    GameUpdate:
      type: object
      properties:
        sport:
          type: string
        playerA:
          type: string
        playerB:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        status:
          type: string
    SettleGameInput:
      type: object
      required: [result]
      properties:
        result:
          type: string
          enum: [player_a_wins, player_b_wins, draw]
    GameListResponse:
      type: object
      required: [games, total, page, limit]
      properties:
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # BETS
    Bet:
      type: object
      required: [id, userId, gameId, outcome, stake, status, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: string
        userId:
          type: integer
        userFullName:
          type: ["string", "null"]
        gameId:
          type: integer
        game:
          $ref: "#/components/schemas/Game"
        outcome:
          type: string
          enum: [player_a_wins, player_b_wins]
        stake:
          type: number
        potentialReturn:
          type: number
        status:
          type: string
          enum: [pending, matched, live, won, lost, refunded, cancelled]
        matchedBetId:
          type: ["integer", "null"]
        createdAt:
          type: string
        settledAt:
          type: ["string", "null"]
...
    Notification:
      type: object
      required: [id, userId, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [bet_matched, near_match, deposit_received, withdrawal_approved, withdrawal_rejected, stream_expiring, bet_won, bet_lost, match_result, bet_refunded, low_balance]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
    NotificationListResponse:
      type: object
      required: [notifications, total, unreadCount, page, limit]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ADMIN STATS
...
    StreamingReport:
      type: object
      required: [period, totalRevenue, totalAccesses, data]
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        totalAccesses:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReportDataPoint"
    BettingReport:
      type: object
      required: [period, brokerageRevenue, totalBetsPlaced, totalBetsMatched, totalBetPool, data]
      properties:
        period:
          type: string
        brokerageRevenue:
          type: number
        totalBetsPlaced:
          type: integer
        totalBetsMatched:
          type: integer
        totalBetPool:
          type: number
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReportDataPoint"
    WalletReport:
      type: object
      required: [period, totalDeposits, totalWithdrawals, netFlow, data]
      properties:
        period:
          type: string
        totalDeposits:
          type: number
        totalWithdrawals:
          type: number
        netFlow:
          type: number
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReportDataPoint"
    ReportDataPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
        secondary:
          type: ["number", "null"]